requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.7.0
pydantic==2.5.0
scipy==1.11.4
scikit-learn==1.3.2
openai>=1.10.0
python-dotenv>=1.0.0
